pyproject.toml
[project]
name = "chopsticks"
version = "0.1.0"
description = "Ceph stress testing framework using Locust"
readme = "README.md"
authors = [
    {name = "Utkarsh Bhatt", email = "utkarsh.bhatt@canonical.com"}
]
requires-python = ">=3.12"
dependencies = [
    "locust>=2.20.0",
    "pyyaml>=6.0",
    "boto3>=1.34.0",
]

[project.scripts]
chopsticks = "chopsticks.cli:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.19.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
